~Nose is the new eyes~
The program now instead of detecting eyes optionally, it detects noses.
Another new feature is the use of CLAHE instead of just the histogram equalize. Which give the detector more robustness agains hard light conditions.
Plus another the facedetectioncam.py script was fixed so the speed velocity is not too high and every capture taken has a different name.
'histogrameq.py' script added for testing different values of equalizers and to compare them.